Assignment 1
This was a remake of the popular Scrabble game using OOP. The assignment instructions are labelled "/Assignment 1/Assignment1_Instructions.pdf". The files that I coded were Tile.java and Scrabble.java. There is also a file that has the full list of every possible Scrabble word so that the user may guess any word they so wish.